Use `usize` instead of `u64` in some structs
This is more ergonomic now that we use `usize` for the Element Size everywhere.
Some `cargo fmt` changes, as usual.

+++ tools/src/matroska_info.rs
@@ -28,7 +28,7 @@ pub enum InfoError {
     #[error(display = "unexpected element: {}", _0)]
     UnexpectedElement(String),
     #[error(display = "offset {:X?}: got unknown element: {:X?} {:#?}", _0, _1, _2)]
-    UnknownElement(usize, u64, Option<u64>),
+    UnknownElement(usize, u64, Option<usize>),
     #[error(display = "failed parsing: {}", _0)]
     Parse(String),
     #[error(display = "could not read the file: {}", _0)]
